refactor(siteImport): split import guard errors into errors.ts
types.ts crossed the 700-line module size budget; the five typed error
classes are their own responsibility and move to errors.ts. All external
consumers import through the @core/siteImport barrel, unchanged.
